--- a/lib/wallet/api/http/Cardano/Wallet/Api/Http/Shelley/Server.hs
+++ b/lib/wallet/api/http/Cardano/Wallet/Api/Http/Shelley/Server.hs
@@ -166,7 +166,7 @@ import Cardano.Mnemonic
 import Cardano.Pool.Types
     ( PoolId )
 import Cardano.Tx.Balance.Internal.CoinSelection
-    ( SelectionOf (..), SelectionStrategy (..) )
+    ( SelectionOf (..) )
 import Cardano.Wallet
     ( BuiltTx (..)
     , DelegationFee (feePercentiles)
@@ -680,8 +680,6 @@ import qualified Cardano.Wallet.Primitive.Types.Tx.TxMeta as W
     ( Direction (Incoming, Outgoing), TxMeta )
 import qualified Cardano.Wallet.Primitive.Types.Tx.TxOut as TxOut
 import qualified Cardano.Wallet.Primitive.Types.UTxO as UTxO
-import qualified Cardano.Wallet.Primitive.Types.UTxOIndex.Internal as UTxOIndex
-import qualified Cardano.Wallet.Primitive.Types.UTxOSelection as UTxOSelection
 import qualified Cardano.Wallet.Read as Read
 import qualified Cardano.Wallet.Registry as Registry
 import qualified Cardano.Wallet.Write.ProtocolParameters as Write
@@ -2141,23 +2139,22 @@ postTransactionOld
         , BoundedAddressLength k
         , HasDelegation s
         , WalletFlavor s n k
+        , IsOurs s RewardAccount
         )
     => ctx
     -> ArgGenChange s
     -> ApiT WalletId
     -> PostTransactionOldData n
     -> Handler (ApiTransaction n)
-postTransactionOld ctx@ApiLayer{..} genChange (ApiT wid) body = do
-    let pwd = coerce $ body ^. #passphrase . #getApiT
+postTransactionOld ctx@ApiLayer{..} argGenChange (ApiT wid) body = do
     let outs = addressAmountToTxOut <$> body ^. #payments
     let md = body ^? #metadata . traverse . #txMetadataWithSchema_metadata
     let mTTL = body ^? #timeToLive . traverse . #getQuantity
-    mkRwdAcct <- case body ^. #withdrawal of
-        Nothing -> pure selfRewardAccountBuilder
-        Just w -> either liftE pure $ shelleyOnlyRewardAccountBuilder @s @_ @n w
     withWorkerCtx ctx wid liftE liftE $ \wrk -> do
         let db = wrk ^. dbLayer
         era <- liftIO $ NW.currentNodeEra netLayer
+        AnyRecentEra (recentEra :: WriteTx.RecentEra era)
+            <- guardIsRecentEra era
         ttl <- liftIO $ W.transactionExpirySlot ti mTTL
         wdrl <- case body ^. #withdrawal of
             Nothing -> pure NoWithdrawal
@@ -2169,59 +2166,40 @@ postTransactionOld ctx@ApiLayer{..} genChange (ApiT wid) body = do
                 , txMetadata = md
                 , txValidityInterval = (Nothing, ttl)
                 }
-        (sel, tx, txMeta, txTime, pp) <- atomicallyWithHandler
-            (ctx ^. walletLocks) (PostTransactionOld wid) $ do
-            (utxoAvailable, wallet, pendingTxs) <-
-                liftHandler $ W.readWalletUTxOIndex @_ @s @k wrk wid
-            let selectAssetsParams = W.SelectAssetsParams
-                    { outputs = F.toList outs
-                    , pendingTxs
-                    , randomSeed = Nothing
-                    , txContext = txCtx
-                    , utxoAvailableForInputs =
-                        UTxOSelection.fromIndex utxoAvailable
-                    , utxoAvailableForCollateral =
-                        UTxOIndex.toMap utxoAvailable
-                    , wallet
-                    , selectionStrategy = SelectionStrategyOptimal
-                    }
-            pp <- liftIO $ NW.currentProtocolParameters netLayer
-            sel <- liftHandler $ W.selectAssets @_ @s @k @'CredFromKeyK
-                    (MsgWallet . W.MsgBalanceTx >$< wrk ^. logger)
-                    txLayer era pp selectAssetsParams
-                $ const Prelude.id
-            sel' <- liftHandler
-                $ W.assignChangeAddressesAndUpdateDb wrk wid genChange sel
-            (tx, txMeta, txTime, sealedTx) <- liftHandler
-                $ W.buildAndSignTransaction @_ @s @k
-                    wrk wid era mkRwdAcct pwd txCtx sel'
-            liftHandler $ W.submitTx (wrk ^. logger) db netLayer wid
-                BuiltTx
-                    { builtTx = tx
-                    , builtTxMeta = txMeta
-                    , builtSealedTx = sealedTx
-                    }
-            pure (sel, tx, txMeta, txTime, pp)
+        (BuiltTx{..}, txTime) <- atomicallyWithHandler
+            (ctx ^. walletLocks) (PostTransactionOld wid) $ liftIO $
+                W.buildSignSubmitTransaction @k @s @n
+                    ti
+                    db
+                    netLayer
+                    txLayer
+                    (coerce $ getApiT $ body ^. #passphrase)
+                    wid
+                    (W.defaultChangeAddressGen argGenChange (Proxy @k))
+                    (AnyRecentEra recentEra)
+                    (PreSelection $ NE.toList outs)
+                    txCtx
+
+        pp <- liftIO $ NW.currentProtocolParameters netLayer
         mkApiTransaction
             (timeInterpreter netLayer)
             wrk wid
             #pendingSince
             MkApiTransactionParams
-                { txId = tx ^. #txId
-                , txFee = tx ^. #fee
-                , txInputs = NE.toList $ second Just <$> sel ^. #inputs
-                -- TODO: ADP-957:
+                { txId = builtTx ^. #txId
+                , txFee = builtTx ^. #fee
+                , txInputs = builtTx ^. #resolvedInputs
                 , txCollateralInputs = []
-                , txOutputs = tx ^. #outputs
-                , txCollateralOutput = tx ^. #collateralOutput
-                , txWithdrawals = tx ^. #withdrawals
-                , txMeta
-                , txMetadata = tx ^. #metadata
+                , txOutputs = builtTx ^. #outputs
+                , txCollateralOutput = builtTx ^. #collateralOutput
+                , txWithdrawals = builtTx ^. #withdrawals
+                , txMeta = builtTxMeta
+                , txMetadata = builtTx ^. #metadata
                 , txTime
-                , txScriptValidity = tx ^. #scriptValidity
+                , txScriptValidity = builtTx ^. #scriptValidity
                 , txDeposit = W.stakeKeyDeposit pp
                 , txMetadataSchema = TxMetadataDetailedSchema
-                , txCBOR = tx ^. #txCBOR
+                , txCBOR = builtTx ^. #txCBOR
                 }
   where
     ti :: TimeInterpreter (ExceptT PastHorizonException IO)
@@ -2407,11 +2385,7 @@ postTransactionFeeOld ctx@ApiLayer{..} (ApiT walletId) body = do
     -- In the context of a mainnet `minfeeA` value of 44 lovelace/byte the
     -- padding is negligible - less than 1/1000 ada.
     padding :: Quantity "byte" Word
-    padding = Quantity $ 20 + tmpExtraPadding
-      where
-        -- TODO [ADP-2268] This padding should be droppable once postTransaction
-        -- also relies on balanceTx.
-        tmpExtraPadding = 5
+    padding = Quantity 20
 
 constructTransaction
     :: forall (n :: NetworkDiscriminant)