--- a/lib/wallet/integration/src/Test/Integration/Scenario/API/Byron/Transactions.hs
+++ b/lib/wallet/integration/src/Test/Integration/Scenario/API/Byron/Transactions.hs
@@ -326,18 +326,8 @@ spec = describe "BYRON_TRANSACTIONS" $ do
             , expectResponseCode HTTP.status202
             ]
 
-        -- TODO [ADP-2268] This padding should be droppable once postTransaction
-        -- also relies on balanceTx.
-        --
-        -- We need it in the meantime as the new implementation is using a lower
-        -- upper bound for the length of Icarus change addresses.
-        let tmpIcarusPadding =
-                if name == "Icarus wallet"
-                then 3_500 -- 35 bytes * 100 lovelace/byte
-                else 0
         let feeEstMin = getFromResponse (#estimatedMin . #getQuantity) rFeeEst
         let feeEstMax = getFromResponse (#estimatedMax . #getQuantity) rFeeEst
-                + tmpIcarusPadding
 
         r <- postTx @n ctx
             (wByron, Link.createTransactionOld @'Byron, fixturePassphrase)