--- a/lib/wallet/integration/src/Test/Integration/Scenario/API/Shelley/Transactions.hs
+++ b/lib/wallet/integration/src/Test/Integration/Scenario/API/Shelley/Transactions.hs
@@ -106,6 +106,7 @@ import Test.Integration.Framework.DSL
     , Headers (..)
     , Payload (..)
     , between
+    , computeApiCoinSelectionFee
     , counterexample
     , decodeErrorInfo
     , emptyRandomWallet
@@ -1050,7 +1051,7 @@ spec = describe "SHELLEY_TRANSACTIONS" $ do
                 , txMetadata =
                     Nothing
                 , expectedFee =
-                    Quantity 131_400
+                    Quantity 129_500
                 }
             , CreateTransactionWithMetadataTest
                 { testName =
@@ -1060,7 +1061,7 @@ spec = describe "SHELLEY_TRANSACTIONS" $ do
                 , txMetadata =
                     Just $ TxMetadata $ Map.singleton 1 $ TxMetaText "hello"
                 , expectedFee =
-                    Quantity 135_600
+                    Quantity 134_200
                 }
             , CreateTransactionWithMetadataTest
                 { testName =
@@ -1070,7 +1071,7 @@ spec = describe "SHELLEY_TRANSACTIONS" $ do
                 , txMetadata =
                       Just txMetadata_ADP_1005
                 , expectedFee =
-                    Quantity 153_200
+                    Quantity 151_800
                 }
             , CreateTransactionWithMetadataTest
                 { testName =
@@ -1085,7 +1086,7 @@ spec = describe "SHELLEY_TRANSACTIONS" $ do
                 , txMetadata =
                       Just txMetadata_ADP_1005
                 , expectedFee =
-                    Quantity 167_200
+                    Quantity 165_200
                 }
             ]
 
@@ -2471,15 +2472,15 @@ spec = describe "SHELLEY_TRANSACTIONS" $ do
                 (`shouldSatisfy` (not . null))
             ]
 
-        -- The expectation below is disabled until the ADP-2268 is implemented,
-        -- (The fees aren't guaranteed to be the same between the two endpoints
-        -- because the 'selectCoins' endpoint has already been updated to
-        -- use the new coin selection functionality while the 'postTransaction'
-        -- endpoint still uses the old one, which gives a slightly different
-        -- estimation)
-        -- let apiCoinSelection = getFromResponse Prelude.id coinSelectionResponse
-        -- let fee = computeApiCoinSelectionFee apiCoinSelection
-        -- Quantity (fromIntegral (unCoin (fee))) `shouldBe` expectedFee
+
+        let apiCoinSelection = getFromResponse Prelude.id coinSelectionResponse
+        let fee = computeApiCoinSelectionFee apiCoinSelection
+        let withinToleranceTo (Quantity expected) tol
+                = between (Quantity $ expected - tol, Quantity $ expected + tol)
+        liftIO $ withinToleranceTo
+            expectedFee
+            1_000
+            (Quantity (fromIntegral (unCoin (fee))))
 
         -- Next, actually create a transaction and submit it to the network.
         -- This transaction should have a fee that is identical to the fee
@@ -2491,6 +2492,7 @@ spec = describe "SHELLEY_TRANSACTIONS" $ do
                 }|]
         ra <- request @(ApiTransaction n) ctx
             (Link.createTransactionOld @'Shelley wa) Default payload
+
         verify ra
             [ expectSuccess
             , expectResponseCode HTTP.status202
@@ -2499,7 +2501,8 @@ spec = describe "SHELLEY_TRANSACTIONS" $ do
                 #metadata
                 (`shouldBe` detailedMetadata <$> txMetadata)
             , expectField
-                (#fee) (`shouldBe` expectedFee)
+                #fee
+                (withinToleranceTo expectedFee 0)
             ]
 
         eventually "metadata is confirmed in transaction list" $ do