--- a/lib/wallet/integration/src/Test/Integration/Scenario/API/Shelley/Transactions.hs
+++ b/lib/wallet/integration/src/Test/Integration/Scenario/API/Shelley/Transactions.hs
@@ -2458,7 +2458,7 @@ spec = describe "SHELLEY_TRANSACTIONS" $ do
 
         -- First, perform a dry-run selection using the 'selectCoins' endpoint.
         -- This will allow us to confirm that the 'selectCoins' endpoint
-        -- produces a selection whose fee is identical to the selection
+        -- produces a selection whose fee is similar to the selection
         -- produced by the 'postTransaction' endpoint.
         coinSelectionResponse <-
             selectCoinsWith @n @'Shelley ctx wa payments maybeAddTxMetadata
@@ -2477,6 +2477,17 @@ spec = describe "SHELLEY_TRANSACTIONS" $ do
         let fee = computeApiCoinSelectionFee apiCoinSelection
         let withinToleranceTo (Quantity expected) tol
                 = between (Quantity $ expected - tol, Quantity $ expected + tol)
+
+        -- Some tolerance is needed as 'selectCoins' doesn't make room for a
+        -- validity interval / TTL, whereas 'postTransaction' does. We seem to
+        -- need 6 bytes / 600 lovelace, but using a bit more to rule out
+        -- flakiness from variable size encoding of slot numbers.
+        --
+        --
+        -- TODO Consider replacing with golden tests on the unit level
+        -- for fee values.
+        --
+        -- Related to https://input-output.atlassian.net/browse/ADP-2935
         liftIO $ withinToleranceTo
             expectedFee
             1_000
@@ -2502,7 +2513,13 @@ spec = describe "SHELLEY_TRANSACTIONS" $ do
                 (`shouldBe` detailedMetadata <$> txMetadata)
             , expectField
                 #fee
-                (withinToleranceTo expectedFee 0)
+                (withinToleranceTo expectedFee 1_000)
+                -- Rule out flakiness from variable size of slot numbers.
+                --
+                -- TODO Consider replacing with golden tests on the unit level
+                -- for fee values.
+                --
+                -- Related to https://input-output.atlassian.net/browse/ADP-2935
             ]
 
         eventually "metadata is confirmed in transaction list" $ do