- [x] Make the switch to `balanceTx` reuse in `postTransactionOld` leveraging the work that we merged with #3824 
- [x] Adjust some hard-coded fee expectations  

### Comments

### Issue Number

ADP-2268