@Anviking 2023-04-21
I tweaked the latency benchmark locally to get it to run with with 20k UTxOs on a Shelley wallet:

The results doesn't make sense, as the balanceTx seems to be faster, which I can't believe. However, in conjunction with the minor slowdown for [1000 UTxOs](https://github.com/input-output-hk/cardano-wallet/pull/3746#issuecomment-1516432587), I take the lack of massive slowdown for 20k as a good sign.

This PR (with PParam fix added: d01627c553aad0b7cd36857881870138abde665a)
```
        listWallets         - 3737.7 ms
        getWallet           - 2667.6 ms
        getUTxOsStatistics  - 3240.1 ms
        listAddresses       - 8835.0 ms
        listTransactions    - 12287.0 ms
        getTransaction      - 5716.3 ms
        postTransactionFee  - 11244.1 ms
        postTransaction     - 10862.8 ms
        postTransTo5Addrs   - 356.4 ms
        postTransactionMA   - 47.3 ms
        listStakePools      - 18.1 ms
        getNetworkInfo      - 0.1 ms
        listMultiAssets     - 0.7 ms
        getMultiAsset       - 0.5 ms
        postMigrationPlan   - 0.9 ms
        postMigration       - 43.8 ms
```

```
        listWallets         - 5150.4 ms
        getWallet           - 4447.9 ms
        getUTxOsStatistics  - 3868.8 ms
        listAddresses       - 9732.7 ms
        listTransactions    - 12228.7 ms
        getTransaction      - 4694.8 ms
        postTransactionFee  - 12074.4 ms
        postTransaction     - 12544.5 ms
        postTransTo5Addrs   - 370.9 ms
        postTransactionMA   - 33.1 ms
        listStakePools      - 26.2 ms
        getNetworkInfo      - 0.1 ms
        listMultiAssets     - 0.7 ms
        getMultiAsset       - 0.5 ms
        postMigrationPlan   - 1.0 ms
        postMigration       - 44.0 ms
```

master (40de2959d4f10348b6b0f76bd06123890af07e9b):
```
        listWallets         - 4963.6 ms
        getWallet           - 4359.9 ms
        getUTxOsStatistics  - 4483.5 ms
        listAddresses       - 10099.1 ms
        listTransactions    - 12665.0 ms
        getTransaction      - 4121.5 ms
        postTransactionFee  - 11356.2 ms
        postTransaction     - 18466.7 ms
        postTransTo5Addrs   - 262.2 ms
        postTransactionMA   - 16.3 ms
        listStakePools      - 23.5 ms
        getNetworkInfo      - 0.1 ms
        listMultiAssets     - 0.7 ms
        getMultiAsset       - 0.5 ms
        postMigrationPlan   - 1.1 ms
        postMigration       - 44.3 ms
```

```
        listWallets         - 4327.9 ms
        getWallet           - 4949.6 ms
        getUTxOsStatistics  - 3023.1 ms
        listAddresses       - 10116.6 ms
        listTransactions    - 12750.9 ms
        getTransaction      - 4793.2 ms
        postTransactionFee  - 11820.2 ms
        postTransaction     - 20140.9 ms
        postTransTo5Addrs   - 287.3 ms
        postTransactionMA   - 16.3 ms
        listStakePools      - 22.7 ms
        getNetworkInfo      - 0.1 ms
        listMultiAssets     - 0.7 ms
        getMultiAsset       - 0.5 ms
        postMigrationPlan   - 0.9 ms
        postMigration       - 46.7 ms
```